Move tests and fix namespace
